requirements.txt
# Основные зависимости
aiogram==3.3.0
python-dotenv==1.0.0

# Веб-сервер для Railway
aiohttp==3.9.1

# Утилиты
pytz==2023.3
python-telegram-bot==20.7
requests==2.31.0

# Логирование
coloredlogs==15.0.1